BookDetails refactor

+++ AccountM.Domain/Role.AD/Permission.cs
@@ -2,10 +2,10 @@
 {
     public class Permission
     {
-        public long Id { get; private set; }//کد اصلی ;برای دیتابیس به permishioncodeربطی ندارد مثل رابطه بین نام و کد ملی.....
+        public int Id { get; private set; }//کد اصلی ;برای دیتابیس به permishioncodeربطی ندارد مثل رابطه بین نام و کد ملی.....
         public int PermissionCode { get; private set; }//آیدی هر دسترسی ==>میتواند چند دسترسی داشته باشد مدیر معلوم میکند
         public string NamePermission { get; private set; }
-        public long? RoleId { get; private set; }//آیدی هر نقش==>متواند فط یک نقش در سایت داشته باشد : مدیر >ادمین>کتابخان>کاربر
+        public int? RoleId { get; private set; }//آیدی هر نقش==>متواند فط یک نقش در سایت داشته باشد : مدیر >ادمین>کتابخان>کاربر
         public Role Role { get; private set; }
 
         public Permission()
